+++ localflavor/co/forms.py
@@ -1,10 +1,10 @@
 """Colombian-specific form helpers."""
 
-from django.forms.fields import Select
 from django.forms import ValidationError
-from django.forms.fields import RegexField
-from django.core.validators import EMPTY_VALUES
+from django.forms.fields import RegexField, Select
 from django.utils.translation import ugettext_lazy as _
+
+from localflavor.compat import EmptyValueCompatMixin
 
 from .co_departments import DEPARTMENT_CHOICES
 
@@ -16,11 +16,11 @@ class CODepartmentSelect(Select):
         super(CODepartmentSelect, self).__init__(attrs, choices=DEPARTMENT_CHOICES)
 
 
-class RUTField(RegexField):
+class CONITField(EmptyValueCompatMixin, RegexField):
     """
     This field validates a NIT (NUmero de IdentificaciOn Tributaria). A
-    NIT is of the form XXXXXXXXXX-V. The last digit is a check digit. Applies
-    to people and companies.
+    NIT is of the form XXXXXXXXXX-V. The last digit is a check digit. This
+    field can be used for people and companies.
 
     More info:
     http://es.wikipedia.org/wiki/N%C3%BAmero_de_Identificaci%C3%B3n_Tributaria
@@ -33,7 +33,7 @@ class RUTField(RegexField):
     PRIME_PLACES = [3, 7, 13, 17, 19, 23, 29, 37, 41, 43, 47, 53, 59, 67, 71]
 
     def __init__(self, max_length=None, min_length=None, *args, **kwargs):
-        super(RUTField, self).__init__(
+        super(CONITField, self).__init__(
             r'^\d{5,12}-?\d$',
             max_length,
             min_length,
@@ -46,9 +46,9 @@ class RUTField(RegexField):
         Value can be either a string in the format XXXXXXXXXX-Y or
         XXXXXXXXXXY.
         """
-        value = super(RUTField, self).clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        value = super(CONITField, self).clean(value)
+        if value in self.empty_values:
+            return self.empty_value
         value, cd = self._canon(value)
         if self._calc_cd(value) != cd:
             raise ValidationError(self.error_messages['checksum'])